+++ packages/data-sql/src/types/clauses/where/index.ts
@@ -1,3 +1,3 @@
-export * from './conditions.js';
+export * from './condition.js';
 export * from './conditions/index.js';
 export * from './logical.js';
